+++ test/fixtures/custom-loader/foo.js
@@ -1,0 +1,6 @@
+'use strict';
+
+module.exports = function(source) {
+  this.cacheable && this.cacheable();
+  return source + 'foo';
+};